fix: drain _benchmarks/ ruff follow-up findings (543 -> 203)
Extended the T201/B023 per-file-ignore to the top-level benchmarks/** glob
(the prior pattern only covered src/mlframe/**/benchmarks/**, missing the
repo-root benchmarks/ directory) -- 543 -> 380.

Ran `ruff check . --ignore C901 --fix --unsafe-fixes` scoped to the remaining
categories (B007 unused loop vars, RUF059 unused unpacked vars, F841 unused
locals, PERF401/402/403 list-building rewrites, N811/N812/N814 import-casing,
RUF005/013/015/034/046 concatenation/slice/format rewrites, E731/E741 lambda
and ambiguous-name renames, B008/B028 default-arg and stacklevel fixes,
PERF102/203 dict-view and try-in-loop hints, G004 f-string-in-log rewrite) --
380 -> 203. Verified every touched file (116) still parses, then hand-cleaned
2 dangling no-op expression statements the F841 fix left behind (an unused
variable's RHS kept as a bare statement with no assignment) in fs_selectors.py
and h2_bench.py -- removed outright since the expressions have no side effect.

All T201 findings are now resolved via the per-file ignore. DTZ005 (41
findings, naive datetime.now() usage) was deliberately NOT auto-fixed -- its
"safe" fix silently adds tz=timezone.utc, which changes behavior and needs
per-call judgment on whether these bench timestamps are meant to be UTC or
local.

Remaining 203 findings (PERF401 x41, DTZ005 x41, N812 x38, B007 x25,
PERF203 x18, F821 x10, B008 x8, E741 x6, N811 x5, N999 x3, N814 x2, and a
handful of singletons) need per-finding judgment and stay a tracked
follow-up, not folded into this mechanical pass.

+++ src/mlframe/feature_selection/filters/_benchmarks/bench_mdlp_robustness.py
@@ -39,7 +39,7 @@ def _inject_duplicates(x: np.ndarray, y: np.ndarray, dup_rate: float, seed: int)
     row -- inflates apparent n without adding independent (x, y) evidence."""
     rng = np.random.default_rng(seed + 9973)
     n = x.shape[0]
-    n_dup = int(round(dup_rate * n))
+    n_dup = round(dup_rate * n)
     if n_dup <= 0:
         return x, y
     dup_idx = rng.integers(0, n, n_dup)
